requirements.txt
PyQt6>=6.6.0
python-vlc>=3.0.18
mutagen>=1.47.0
librosa>=0.10.1
numpy>=1.24.0
pandas>=2.0.0 